+++ src/FabianBeiner/Todoist/TodoistTasksTrait.php
@@ -3,14 +3,13 @@
  * Todoist PHP API Library
  * An unofficial PHP client library for accessing the official Todoist REST API.
  *
+ * @author  Fabian Beiner <fb@fabianbeiner.de>
  * @author  Balazs Csaba <balazscsaba2006@gmail.com>
- * @license MIT
+ * @license https://opensource.org/licenses/MIT MIT
  * @link    https://github.com/FabianBeiner/Todoist-PHP-API-Library
  */
 
 namespace FabianBeiner\Todoist;
-
-use GuzzleHttp\RequestOptions;
 
 /**
  * Trait TodoistTasksTrait.
@@ -26,13 +25,13 @@ trait TodoistTasksTrait
      */
     public function getAllTasks()
     {
-        $result = $this->client->get('tasks');
+        $result = $this->get('tasks');
 
         $status = $result->getStatusCode();
-        if ($status === 204) {
+        if (204 === $status) {
             return [];
         }
-        if ($status === 200) {
+        if (200 === $status) {
             return json_decode($result->getBody()->getContents());
         }
 
@@ -43,24 +42,22 @@ trait TodoistTasksTrait
      * Create a new task.
      *
      * @param string $content Content of the task.
-     * @param array $options Possibility to add non-required parameters.
+     * @param array  $options Possibility to add non-required parameters.
      *
      * @return array|bool Array with values of the new task, or false on failure.
      */
-    public function createTask($content, array $options = [])
+    public function createTask(string $content, array $options = [])
     {
         if ('' === $content) {
             return false;
         }
 
-        if ($options) {
-            unset($options['content']);
-        }
-        $result = $this->client->post('tasks', [
-            RequestOptions::JSON => array_merge(['content' => trim($content)], $options),
-        ]);
+        unset($options['content']);
+        $data = $this->prepareRequestData(array_merge(['content' => $content], $options));
+        $result = $this->post('tasks', $data);
 
-        if ($result->getStatusCode() === 200) {
+        $status = $result->getStatusCode();
+        if (200 === $status) {
             return json_decode($result->getBody()->getContents());
         }
 
@@ -74,15 +71,16 @@ trait TodoistTasksTrait
      *
      * @return array|bool Array with values of the task, or false on failure.
      */
-    public function getTask($taskId)
+    public function getTask(int $taskId)
     {
-        if ($taskId <= 0 || !filter_var($taskId, FILTER_VALIDATE_INT)) {
+        if (!$this->validateId($taskId)) {
             return false;
         }
 
-        $result = $this->client->get('tasks/' . $taskId);
+        $result = $this->get('tasks/' . $taskId);
 
-        if ($result->getStatusCode() === 200) {
+        $status = $result->getStatusCode();
+        if (200 === $status) {
             return json_decode($result->getBody()->getContents());
         }
 
@@ -92,30 +90,23 @@ trait TodoistTasksTrait
     /**
      * Update a task.
      *
-     * @param int $taskId ID of the task.
+     * @param int    $taskId  ID of the task.
      * @param string $content Content of the task.
-     * @param array $options Possibility to add non-required parameters.
+     * @param array  $options Possibility to add non-required parameters.
      *
-     * @return array|bool Array with values of the new task, or false on failure.
+     * @return bool
      */
-    public function updateTask($taskId, $content, array $options = [])
+    public function updateTask(int $taskId, string $content, array $options = []): bool
     {
         if ('' === $content) {
             return false;
         }
 
-        if ($options) {
-            unset($options['content']);
-        }
-        $result = $this->client->post('tasks/' . $taskId, [
-            RequestOptions::JSON => array_merge(['content' => trim($content)], $options),
-        ]);
+        unset($options['content']);
+        $data = $this->prepareRequestData(array_merge(['content' => $content], $options));
+        $result = $this->post('tasks/' . $taskId, $data);
 
-        if ($result->getStatusCode() === 200) {
-            return json_decode($result->getBody()->getContents());
-        }
-
-        return false;
+        return 204 === $result->getStatusCode();
     }
 
     /**
@@ -125,16 +116,15 @@ trait TodoistTasksTrait
      *
      * @return bool True on success, false on failure.
      */
-    public function closeTask($taskId): bool
+    public function closeTask(int $taskId): bool
     {
-        if ($taskId <= 0 || !filter_var($taskId, FILTER_VALIDATE_INT)) {
+        if (!$this->validateId($taskId)) {
             return false;
         }
 
-        $result = $this->client->get('tasks/' . $taskId . '/close');
-        $status = $result->getStatusCode();
+        $result = $this->get('tasks/' . $taskId . '/close');
 
-        return ($status === 200 || $status === 204);
+        return 204 === $result->getStatusCode();
     }
 
     /**
@@ -144,16 +134,15 @@ trait TodoistTasksTrait
      *
      * @return bool True on success, false on failure.
      */
-    public function reopenTask($taskId): bool
+    public function reopenTask(int $taskId): bool
     {
-        if ($taskId <= 0 || !filter_var($taskId, FILTER_VALIDATE_INT)) {
+        if (!$this->validateId($taskId)) {
             return false;
         }
 
-        $result = $this->client->get('tasks/' . $taskId . '/reopen');
-        $status = $result->getStatusCode();
+        $result = $this->get('tasks/' . $taskId . '/reopen');
 
-        return ($status === 200 || $status === 204);
+        return 204 === $result->getStatusCode();
     }
 
     /**
@@ -163,15 +152,14 @@ trait TodoistTasksTrait
      *
      * @return bool True on success, false on failure.
      */
-    public function deleteTask($taskId): bool
+    public function deleteTask(int $taskId): bool
     {
-        if ($taskId <= 0 || !filter_var($taskId, FILTER_VALIDATE_INT)) {
+        if (!$this->validateId($taskId)) {
             return false;
         }
 
-        $result = $this->client->delete('tasks/' . $taskId);
-        $status = $result->getStatusCode();
+        $result = $this->delete('tasks/' . $taskId);
 
-        return ($status === 200 || $status === 204);
+        return 204 === $result->getStatusCode();
     }
 }