+++ CHANGELOG.md
@@ -1,6 +1,24 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## 0.6.0 - 2018-03-06
+### Changed
+- Improved the code quality overall.
+- Added a helper function from @balazscsaba2006.
+- Enabled GZIP.
+- Requires PHP 7.x.
+- Some more smaller improvements.
+
+## 0.5.0 - 2018-03-02
+### Changed
+- Improved the code of old Traits.
+- Added `TodoistProjectsTrait.php` from @balazscsaba2006.
+- Implemented all the Projects endpoints.
+
+## 0.4.0 - 2017-11-14
+### Added
+- Implemented all the Comments endpoints.
+
 ## 0.3.1 - 2017-11-11
 ### Changed
 - Removed ramsey/uuid.